+++ CHANGELOG.md
@@ -41,6 +41,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Snapshot synchronization could remove committed log entries that not be included in snapshots
   [#167](https://github.com/lerna-stack/akka-entity-replication/issues/167)
   [#PR168](https://github.com/lerna-stack/akka-entity-replication/pull/168)
+- SnapshotStore doesn't reply with SnapshotNotFound sometimes
+  [#182](https://github.com/lerna-stack/akka-entity-replication/issues/182),
+  [#PR183](https://github.com/lerna-stack/akka-entity-replication/pull/183)
 
 ## [v2.1.0] - 2022-03-24
 [v2.1.0]: https://github.com/lerna-stack/akka-entity-replication/compare/v2.0.0...v2.1.0